requirements.txt
flask==2.0.1
flask-cors==3.0.10
pillow==9.0.0
numpy==1.22.0
boto3==1.20.26
requests==2.27.1